Oops everything is changed
Ahahahaha I didn't like that last theme and wasn't feeling it so I changed it. I liked the black and white but there was too much negative space and I don't really like that redish pink color that much but it was my default color. Now I'm exmphasizing the sky and stars and my abilities a lot more than I was before